requirements.txt
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
dnspython==1.16.0
emoji==0.6.0
idna==2.10
pycodestyle==2.6.0
pymongo==3.11.2
pyTelegramBotAPI==3.7.6
requests==2.25.1
toml==0.10.2
urllib3==1.26.2
